requirements.txt
dataclasses==0.6
google-api-python-client==1.12.5
ibis-framework==1.3.0
impyla==0.16.2
sqlalchemy==1.3.18
pymysql==0.10.0
psycopg2-binary==2.8.5
PyYAML==5.3.1
# pyspark==3.0.0
apache-airflow==1.10.11
pandas==1.0.5
pyarrow==0.17.1
google-cloud-bigquery==1.26.1
google-cloud-bigquery-storage==1.0.0
setuptools>=34.0.0
jellyfish==0.8.2
